scripts/requirements.txt
pillow
cairosvg
# For web scraping
requests 
beautifulsoup4
tqdm
cloudscraper